+++ ide/configwin_ihm.ml
@@ -27,6 +27,10 @@
 
 open Configwin_types
 
+let set_help_tip wev = function
+  | None -> ()
+  | Some help -> GtkBase.Widget.Tooltip.set_text wev#as_widget help
+
 let modifiers_to_string m =
   let rec iter m s =
     match m with
@@ -55,7 +59,7 @@ class type widget =
 
 let debug = false
 let dbg s = if debug then Minilib.log s else ()
-
+(*
 (** This class builds a frame with a clist and two buttons :
    one to add items and one to remove the selected items.
    The class takes in parameter a function used to add items and
@@ -71,7 +75,6 @@ class ['a] list_selection_box
     f_color
     (eq : 'a -> 'a -> bool)
     add_function title editable
-    (tt:GData.tooltips)
     =
   let _ = dbg "list_selection_box" in
   let wev = GBin.event_box () in
@@ -94,12 +97,8 @@ class ['a] list_selection_box
 	~titles_show: true
 	~packing: wscroll#add ()
   in
-  let _ =
-    match help_opt with
-      None -> ()
-    | Some help ->
-	tt#set_tip ~text: help ~privat: help wev#coerce
-  in  (* the vbox for the buttons *)
+  let _ = set_help_tip wev help_opt in
+  (* the vbox for the buttons *)
   let vbox_buttons = GPack.vbox () in
   let _ =
     if editable then
@@ -279,10 +278,10 @@ class ['a] list_selection_box
       (* initialize the clist with the listref *)
       self#update !listref
   end;;
-
+*)
 
 (** This class is used to build a box for a string parameter.*)
-class string_param_box param (tt:GData.tooltips) =
+class string_param_box param =
   let _ = dbg "string_param_box" in
   let hbox = GPack.hbox () in
   let wev = GBin.event_box ~packing: (hbox#pack ~expand: false ~padding: 2) () in
@@ -292,12 +291,7 @@ class string_param_box param (tt:GData.tooltips) =
       ~packing: (hbox#pack ~expand: param.string_expand ~padding: 2)
       ()
   in
-  let _ =
-    match param.string_help with
-      None -> ()
-    | Some help ->
-	tt#set_tip ~text: help ~privat: help wev#coerce
-  in
+  let _ = set_help_tip wev param.string_help in
   let _ = we#set_text (param.string_to_string param.string_value) in
 
   object (self)
@@ -316,17 +310,12 @@ class string_param_box param (tt:GData.tooltips) =
   end ;;
 
 (** This class is used to build a box for a combo parameter.*)
-class combo_param_box param (tt:GData.tooltips) =
+class combo_param_box param =
     let _ = dbg "combo_param_box" in
     let hbox = GPack.hbox () in
     let wev = GBin.event_box ~packing: (hbox#pack ~expand: false ~padding: 2) () in
     let _wl = GMisc.label ~text: param.combo_label ~packing: wev#add () in
-    let _ =
-      match param.combo_help with
-	  None -> ()
-	| Some help ->
-	    tt#set_tip ~text: help ~privat: help wev#coerce
-    in
+    let _ = set_help_tip wev param.combo_help in
     let get_value = if not param.combo_new_allowed then
       let wc = GEdit.combo_box_text
 	~strings: param.combo_choices
@@ -341,13 +330,13 @@ class combo_param_box param (tt:GData.tooltips) =
 	fun () -> match GEdit.text_combo_get_active wc with |None -> "" |Some s -> s
     else
       let (wc,_) = GEdit.combo_box_entry_text
-	~strings: param.combo_choices
-	~packing: (hbox#pack ~expand: param.combo_expand ~padding: 2)
-	()
+       ~strings: param.combo_choices
+       ~packing: (hbox#pack ~expand: param.combo_expand ~padding: 2)
+       ()
       in
       let _ = wc#entry#set_editable param.combo_editable in
       let _ = wc#entry#set_text param.combo_value in
-	fun () -> wc#entry#text
+       fun () -> wc#entry#text
     in
 object (self)
 
@@ -365,7 +354,7 @@ object (self)
 end ;;
 
 (** Class used to pack a custom box. *)
-class custom_param_box param (tt:GData.tooltips) =
+class custom_param_box param =
   let _ = dbg "custom_param_box" in
   let top =
     match param.custom_framed with
@@ -381,7 +370,7 @@ class custom_param_box param (tt:GData.tooltips) =
   end
 
 (** This class is used to build a box for a text parameter.*)
-class text_param_box param (tt:GData.tooltips) =
+class text_param_box param =
   let _ = dbg "text_param_box" in
   let wf = GBin.frame ~label: param.string_label ~height: 100 () in
   let wev = GBin.event_box ~packing: wf#add () in
@@ -395,12 +384,7 @@ class text_param_box param (tt:GData.tooltips) =
       ~packing: wscroll#add
       ()
   in
-  let _ =
-    match param.string_help with
-      None -> ()
-    | Some help ->
-	tt#set_tip ~text: help ~privat: help wev#coerce
-  in
+  let _ = set_help_tip wev param.string_help in
   let _ = dbg "text_param_box: buffer creation" in
   let buffer = GText.buffer () in
 
@@ -427,17 +411,13 @@ class text_param_box param (tt:GData.tooltips) =
   end ;;
 
 (** This class is used to build a box for a boolean parameter.*)
-class bool_param_box param (tt:GData.tooltips) =
+class bool_param_box param =
   let _ = dbg "bool_param_box" in
   let wchk = GButton.check_button
       ~label: param.bool_label
       ()
   in
-  let _ =
-    match param.bool_help with
-      None -> ()
-    | Some help -> tt#set_tip ~text: help ~privat: help wchk#coerce
-  in
+  let _ = set_help_tip wchk param.bool_help in
   let _ = wchk#set_active param.bool_value in
   let _ = wchk#misc#set_sensitive param.bool_editable in
 
@@ -471,14 +451,7 @@ class modifiers_param_box param =
                                  else value := List.filter ((<>) modifier) !value)))
             param.md_allow
   in
-  let _ =
-    match param.md_help with
-      None -> ()
-    | Some help ->
-       let tooltips = GData.tooltips () in
-       ignore (hbox#connect#destroy ~callback: tooltips#destroy);
-       tooltips#set_tip wev#coerce ~text: help ~privat: help
-  in
+  let _ = set_help_tip wev param.md_help in
   object (self)
 
     (** This method returns the main box ready to be packed. *)
@@ -493,9 +466,9 @@ class modifiers_param_box param =
       else
 	()
   end ;;
-
+(*
 (** This class is used to build a box for a parameter whose values are a list.*)
-class ['a] list_param_box (param : 'a list_param) (tt:GData.tooltips) =
+class ['a] list_param_box (param : 'a list_param) =
   let _ = dbg "list_param_box" in
   let listref = ref param.list_value in
   let frame_selection = new list_selection_box
@@ -520,9 +493,10 @@ class ['a] list_param_box (param : 'a list_param) (tt:GData.tooltips) =
       param.list_f_apply !listref ;
       param.list_value <- !listref
   end ;;
+*)
 
 (** This class creates a configuration box from a configuration structure *)
-class configuration_box (tt : GData.tooltips) conf_struct =
+class configuration_box conf_struct =
 
   let main_box = GPack.hbox () in
 
@@ -553,27 +527,27 @@ class configuration_box (tt : GData.tooltips) conf_struct =
 
   let make_param (main_box : #GPack.box) = function
   | String_param p ->
-    let box = new string_param_box p tt in
+    let box = new string_param_box p in
     let _ = main_box#pack ~expand: false ~padding: 2 box#box in
     box
   | Combo_param p ->
-    let box = new combo_param_box p tt in
+    let box = new combo_param_box p in
     let _ = main_box#pack ~expand: false ~padding: 2 box#box in
     box
   | Text_param p ->
-    let box = new text_param_box p tt in
+    let box = new text_param_box p in
     let _ = main_box#pack ~expand: p.string_expand ~padding: 2 box#box in
     box
   | Bool_param p ->
-    let box = new bool_param_box p tt in
+    let box = new bool_param_box p in
     let _ = main_box#pack ~expand: false ~padding: 2 box#box in
     box
   | List_param f ->
-    let box = f tt in
+    let box = f () in
     let _ = main_box#pack ~expand: true ~padding: 2 box#box in
     box
   | Custom_param p ->
-    let box = new custom_param_box p tt in
+    let box = new custom_param_box p in
     let _ = main_box#pack ~expand: p.custom_expand ~padding: 2 box#box in
     box
   | Modifiers_param p ->
@@ -684,11 +658,9 @@ let edit ?(with_apply=true)
     ?parent ?height ?width
     ()
   in
-  let tooltips = GData.tooltips () in
+  let config_box = new configuration_box conf_struct in
 
-  let config_box = new configuration_box tooltips conf_struct in
-
-  let _ = dialog#vbox#add config_box#box#coerce in
+  let _ = dialog#vbox#pack ~expand:true config_box#box#coerce in
 
   if with_apply then
     dialog#add_button Configwin_messages.mApply `APPLY;
@@ -697,7 +669,6 @@ let edit ?(with_apply=true)
   dialog#add_button Configwin_messages.mCancel `CANCEL;
 
   let destroy () =
-    tooltips#destroy () ;
     dialog#destroy ();
   in
   let rec iter rep =
@@ -714,10 +685,12 @@ let edit ?(with_apply=true)
   in
     iter Return_cancel
 
+(*
 let edit_string l s =
   match GToolbox.input_string ~title: l ~text: s Configwin_messages.mValue with
     None -> s
   | Some s2 -> s2
+*)
 
 (** Create a string param. *)
 let string ?(editable=true) ?(expand=true) ?help ?(f=(fun _ -> ())) label v =
@@ -744,6 +717,7 @@ let bool ?(editable=true) ?help ?(f=(fun _ -> ())) label v =
       bool_f_apply = f ;
     }
 
+(*
 (** Create a list param. *)
 let list ?(editable=true) ?help
     ?(f=(fun (_:'a list) -> ()))
@@ -753,7 +727,7 @@ let list ?(editable=true) ?help
     ?titles ?(color=(fun (_:'a) -> (None : string option)))
     label (f_strings : 'a -> string list) v =
   List_param
-    (fun tt ->
+    (fun () ->
 	new list_param_box
 	  {
 	    list_label = label ;
@@ -768,7 +742,6 @@ let list ?(editable=true) ?help
 	    list_f_add = add ;
 	    list_f_apply = f ;
 	  }
-	  tt
     )
 
 (** Create a strings param. *)
@@ -777,6 +750,7 @@ let strings ?(editable=true) ?help
     ?(eq=Pervasives.(=))
     ?(add=(fun () -> [])) label v =
   list ~editable ?help ~f ~eq ~edit: (edit_string label) ~add label (fun s -> [s]) v
+*)
 
 (** Create a combo param. *)
 let combo ?(editable=true) ?(expand=true) ?help ?(f=(fun _ -> ()))